+++ src/ebonite/ext/ext_loader.py
@@ -57,10 +57,13 @@ class ExtensionLoader:
         Extension('ebonite.ext.pandas', ['pandas'], False),
         Extension('ebonite.ext.sklearn', ['sklearn'], False),
         Extension('ebonite.ext.tensorflow', ['tensorflow'], False),
-        Extension('ebonite.ext.flask', ['flask', 'flaskgger']),
+        Extension('ebonite.ext.torch', ['torch'], False),
+        Extension('ebonite.ext.catboost', ['catboost'], False),
+        Extension('ebonite.ext.flask', ['flask', 'flasgger']),
         Extension('ebonite.ext.sqlalchemy', ['sqlalchemy']),
         Extension('ebonite.ext.s3', ['boto3']),
         Extension('ebonite.ext.imageio', ['imageio']),
+        Extension('ebonite.ext.lightgbm', ['lightgbm'], False),
         Extension('ebonite.ext.xgboost', ['xgboost'], False)
     )