+++ src/ebonite/ext/lightgbm/__init__.py
@@ -1,0 +1,4 @@
+from .dataset import LightGBMDatasetHook, LightGBMDatasetType
+from .model import LightGBMModelHook, LightGBMModelWrapper
+
+__all__ = ['LightGBMModelWrapper', 'LightGBMModelHook', 'LightGBMDatasetHook', 'LightGBMDatasetType']