+++ src/ebonite/ext/torch/__init__.py
@@ -1,0 +1,4 @@
+from .dataset import TorchTensorDatasetType, TorchTensorHook
+from .model import TorchModelHook, TorchModelWrapper
+
+__all__ = ['TorchTensorDatasetType', 'TorchTensorHook', 'TorchModelHook', 'TorchModelWrapper']